consolidate chain syscalls into network
