+++ sdk/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod actor;
-pub mod chain;
 pub mod crypto;
 pub mod debug;
 pub mod error;
