+++ sdk/src/network.rs
@@ -3,6 +3,9 @@ use std::convert::TryInto;
 use fvm_shared::clock::ChainEpoch;
 use fvm_shared::econ::TokenAmount;
 use fvm_shared::version::NetworkVersion;
+use cid::Cid;
+use fvm_shared::error::ErrorNumber;
+use fvm_shared::MAX_CID_LEN;
 
 use crate::sys;
 use crate::vm::INVOCATION_CONTEXT;
@@ -31,5 +34,21 @@ pub fn total_fil_circ_supply() -> TokenAmount {
         sys::network::total_fil_circ_supply()
             .expect("failed to get circulating supply")
             .into()
+    }
+}
+
+pub fn tipset_timestamp() -> u64 {
+    unsafe { sys::network::tipset_timestamp() }.expect("failed to get timestamp")
+}
+
+pub fn tipset_cid(epoch: i64) -> Option<Cid> {
+    let mut buf = [0u8; MAX_CID_LEN];
+
+    unsafe {
+        match sys::network::tipset_cid(epoch, buf.as_mut_ptr(), MAX_CID_LEN as u32) {
+            Ok(len) => Some(Cid::read_bytes(&buf[..len as usize]).expect("invalid cid")),
+            Err(ErrorNumber::NotFound) => None,
+            Err(other) => panic!("unexpected cid resolution failure: {}", other),
+        }
     }
 }
